requirements.txt
crewai>=0.1.0
playwright>=1.40.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2 